Refactored inner enum ArrayType to its own file: PooledArrayType
